# quads/swagger.yaml
openapi: 3.0.0
info:
  version: '3.0.0'
  title: 'QUADS'
  description: 'QUADS automates the future scheduling, end-to-end provisioning and delivery of bare-metal servers and networks.'
  license:
    name: GPLv3
servers:
  - url: 'https://quads.example.com/api/v3/'
paths:
...
  /schedules/:
    get:
      summary: 'Returns a list with all the defined schedules'
      tags:
        - schedules
      responses:
        '200':
          description: 'Schedule'
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Add a new schedule
      tags:
        - schedules
      requestBody:
        description: Body request with *schedule* details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
          application/xml:
            schema:
              $ref: '#/components/schemas/Schedule'
          text/plain:
            schema:
              type: string
      responses:
        '201':
          description: Created
  /schedules/{schedule_id}:
    get:
      summary: 'Returns a schedule by id'
      tags:
        - schedules
      parameters:
        - name: schedule_id
          in: path
          description: Schedule ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Schedule'
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      summary: 'Update an existing schedule'
      tags:
        - schedules
      parameters:
        - name: schedule_id
          in: path
          description: Schedule ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Schedule'
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      description: Delete schedule by id
      tags:
        - schedules
      parameters:
        - name: schedule_id
          in: path
          description: Schedule ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Deleted Schedule'
  /schedules/current:
    get:
      summary: 'Returns all current schedules'
      tags:
        - schedules
      parameters:
        - name: schedule_id
          in: path
          description: Schedule ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Schedule'
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /schedules/future:
    get:
      summary: 'Returns all future schedules'
      tags:
        - schedules
      parameters:
        - name: schedule_id
          in: path
          description: Schedule ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Schedule'
          headers:
            x-next:
              description: A link to the next page of responses
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Schedule:
      type: object
      properties:
        start:
          type: string
          example: 2022-02-02
        end:
          type: string
          example: 2022-04-02
        build_start:
          type: string
          example: 2022-02-02
        build_end:
          type: string
          example: 2022-04-02
        assignment_id:
          type: integer
          example: 12
        host_id:
          type: integer
          example: 1
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer